simplify vm-babel-plugins for direct consumption
